Move `createReferenceNode`
