+++ packages/seroval/src/core/tree/sync.ts
@@ -15,7 +15,6 @@ import {
 import {
   createBigIntTypedArrayNode,
   createTypedArrayNode,
-  createReferenceNode,
   createDataViewNode,
   createSymbolNode,
   createFunctionNode,
@@ -56,6 +55,7 @@ import {
   createDateNode,
   createIndexedValueNode,
   createNumberNode,
+  createReferenceNode,
   createRegExpNode,
   createStringNode,
 } from '../base-primitives';
